requirements.txt
fastapi
uvicorn[standard]
python-multipart
SQLAlchemy
pydantic
faiss-cpu
sentence-transformers
PyPDF2
numpy
openai
python-dotenv
